dev_requirements.txt
# Requirements for doing development

# Keep pip unpinned and always update
pip

# For Sphinx documentation
Sphinx==1.6.6
sphinx-rtd-theme==0.2.4
sphinx-click==1.0.2

# For testing. Do *not* put these into test_requirements.txt since we don't want
# to install tox in the environment created by tox. It won't get used.
tox==2.9.1
tox-pyenv==1.1.0
